pipelines/rj_iplanrio__alertario_previsao_24h/pyproject.toml
[project]
name = "rj_iplanrio__alertario_previsao_24h"
version = "0.1.0"
description = "Pipeline previsão meteorológica AlertaRio 24h da secretaria iplanrio"
dependencies = [
    "defusedxml>=0.7.1",
    "google-cloud-bigquery>=3.25.0",
    "basedosdados>=2.0.0b23",
    "pandas>=2.3.1",
    "requests>=2.32.4",
]
